Fixed the bug
When there were two characters on the main screen (one in the first position and one in any other position), removing the character from the non-first position and then removing the first one caused an issue. Regardless of which character was selected to spawn in the first position, the image of the last removed character would appear. The Blueprint class was spawned correctly, but the parameters (such as which character is on the main screen) were saved incorrectly.